Possible fix for chemostats in reaction builder
When chemostats were added using the reaction builder, an
InvalidComponentException would be raised, as the chemostat is created
as a Ce component, not an Se component. Also, the anchor appears to be
created incorrectly because the code thinks the chemostat is involved in
one more reaction than it actually is. The changes here appear to
resolve these issues.